refactor: extract `Drug` class from `pharmacy.ts` file
