+++ drug.ts
@@ -1,0 +1,7 @@
+export class Drug {
+  constructor(
+    public name: string,
+    public expiresIn: number,
+    public benefit: number
+  ) {}
+}
