+++ index.ts
@@ -1,5 +1,7 @@
-import { Drug, Pharmacy } from "./pharmacy";
 import fs from "fs";
+
+import { Drug } from "./drug";
+import { Pharmacy } from "./pharmacy";
 
 const drugs = [
   new Drug("Doliprane", 20, 30),
